requirements.txt
alembic==1.7.7
anyio==3.5.0
asgiref==3.5.0
attrs==21.4.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.2
cryptography==36.0.2
databases==0.5.5
fastapi==0.75.2
FastAPI-SQLAlchemy==0.2.1
greenlet==1.1.2
h11==0.13.0
idna==3.3
iniconfig==1.1.1
Mako==1.2.0
MarkupSafe==2.1.1
packaging==21.3
pluggy==1.0.0
psycopg2==2.9.3
py==1.11.0
pycparser==2.21
pydantic==1.9.0
pyparsing==3.0.8
pytest==7.1.2
python-dotenv==0.20.0
requests==2.27.1
sniffio==1.2.0
SQLAlchemy==1.4.35
sqlalchemy2-stubs==0.0.2a22
sqlmodel==0.0.6
starlette==0.17.1
tomli==2.0.1
typing_extensions==4.2.0
urllib3==1.26.9
uvicorn==0.17.6
